Update code in Gen_network.py and gen_route.py to generate 10x10 network.

+++ MiniVnet_cpp/globalConst.h
@@ -33,7 +33,7 @@ extern uint8_t _CHOOSE_CAR_OPTION;
 extern uint8_t _TOP_N_CONGESTED;
 extern uint8_t _ITERATION_NUM;
 extern double _NOW_SIMU_TIME;
-extern double _CAR_TIME_ERROR;	// Time error for triggering cars to reroute
+extern int _CAR_TIME_ERROR;	// Time error for triggering cars to reroute
 
 extern std::string _RANDOM_SEED;
 extern std::string _ARRIVAL_RATE;
